Move `WordSplitter` and `split_words` to new `word_splitters` module
This moves all “word splitting” or “hyphenation” functionality into
its own module. The three main traits now all have their own module,
and the `core` module is much slimmer.
